requirement.txt
streamlit
torch
torchvision
Pillow
matplotlib
optuna